extract SegmentedControls
